+++ server/example/org.eclipse.glsp.ide.workflow.editor/META-INF/MANIFEST.MF
@@ -3,8 +3,9 @@ Bundle-ManifestVersion: 2
 Bundle-Name: Workflow Editor
 Bundle-SymbolicName: org.eclipse.glsp.ide.workflow.editor;singleton:=true
 Bundle-Version: 1.1.0.qualifier
-Automatic-Module-Name: com.eclipsesource.glsp.example.workflow
-Bundle-RequiredExecutionEnvironment: JavaSE-11
+Bundle-ClassPath: .
+Bundle-Activator: org.eclipse.glsp.ide.workflow.editor.Activator
+Bundle-Vendor: Eclipse GLSP
 Require-Bundle: org.eclipse.elk.alg.common;bundle-version="0.8.1",
  org.eclipse.elk.alg.layered;bundle-version="0.8.1",
  org.eclipse.elk.graph.text;bundle-version="0.8.1",
@@ -19,12 +20,12 @@ Require-Bundle: org.eclipse.elk.alg.common;bundle-version="0.8.1",
  org.eclipse.elk.graph;bundle-version="0.8.1",
  org.eclipse.lsp4j.websocket;bundle-version="0.11.0",
  org.eclipse.e4.core.contexts;bundle-version="1.8.400",
- org.eclipse.jetty.server;bundle-version="[10.0.0,11.0.0)"
-Bundle-ClassPath: .
-Eclipse-BundleShape: dir
-Bundle-ActivationPolicy: lazy
-Bundle-Activator: org.eclipse.glsp.ide.workflow.editor.Activator
-Bundle-Vendor: Eclipse GLSP
+ org.eclipse.jetty.server;bundle-version="[10.0.0,11.0.0)",
+ org.eclipse.e4.ui.model.workbench;bundle-version="2.2.300.v20221024-2137"
 Import-Package: javax.servlet;version="[4.0.0,5.0.0)",
  javax.websocket;version="1.1.0",
  org.apache.logging.log4j;version="2.19.0"
+Eclipse-BundleShape: dir
+Automatic-Module-Name: com.eclipsesource.glsp.example.workflow
+Bundle-ActivationPolicy: lazy
+Bundle-RequiredExecutionEnvironment: JavaSE-11
