+++ server/example/org.eclipse.glsp.ide.workflow.editor/build.properties
@@ -1,6 +1,7 @@
-source.. = src/
 bin.includes = META-INF/,\
                .,\
                plugin.xml,\
                diagram/,\
-               icons/
+               icons/,\
+               fragment.e4xmi
+source.. = src/
