+++ server/example/org.eclipse.glsp.ide.workflow.editor/fragment.e4xmi
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="ASCII"?>
+<fragment:ModelFragments xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:basic="http://www.eclipse.org/ui/2010/UIModel/application/descriptor/basic" xmlns:fragment="http://www.eclipse.org/ui/2010/UIModel/fragment" xmlns:menu="http://www.eclipse.org/ui/2010/UIModel/application/ui/menu" xmi:id="_Lo1j0CVVEe6BWNnGTQlFsQ">
+  <fragments xsi:type="fragment:StringModelFragment" xmi:id="_MGE5cCVVEe6BWNnGTQlFsQ" featurename="descriptors" parentElementId="org.eclipse.e4.legacy.ide.application">
+    <elements xsi:type="basic:PartDescriptor" xmi:id="_OELk4CVVEe6BWNnGTQlFsQ" elementId="org.eclipse.glsp.ide.workflow.editor.partdescriptor.glspEditor" label="GLSP Viewer" allowMultiple="true" closeable="true" contributionURI="bundleclass://org.eclipse.glsp.ide.editor/org.eclipse.glsp.ide.editor.ui.GLSPDiagramPart">
+      <menus xsi:type="menu:PopupMenu" xmi:id="_Uj4zQCVVEe6BWNnGTQlFsQ" elementId="org.eclipse.glsp.ide.workflow.editor.popupmenu"/>
+      <toolbar xmi:id="_Sw82ACVVEe6BWNnGTQlFsQ" elementId="org.eclipse.glsp.ide.workflow.editor.toolbar.0"/>
+    </elements>
+  </fragments>
+</fragment:ModelFragments>
