+++ server/plugins/org.eclipse.glsp.ide.editor/META-INF/MANIFEST.MF
@@ -28,7 +28,12 @@ Require-Bundle: org.eclipse.ui.ide;bundle-version="3.18.0",
  org.eclipse.jetty.util;bundle-version="[10.0.0,11.0.0)",
  org.eclipse.jetty.servlet;bundle-version="[10.0.0,11.0.0)",
  org.eclipse.jetty.websocket.javax.server;bundle-version="[10.0.0,11.0.0)",
- org.eclipse.jface.text;bundle-version="3.0.0"
+ org.eclipse.jface.text;bundle-version="3.0.0",
+ org.eclipse.e4.ui.di,
+ org.eclipse.e4.ui.model.workbench,
+ org.eclipse.e4.ui.workbench,
+ org.eclipse.e4.core.commands,
+ org.eclipse.e4.core.di.annotations
 Bundle-ClassPath: .
 Export-Package: org.eclipse.glsp.ide.editor,
  org.eclipse.glsp.ide.editor.actions,
@@ -42,7 +47,8 @@ Export-Package: org.eclipse.glsp.ide.editor,
  org.eclipse.glsp.ide.editor.internal.actions,
  org.eclipse.glsp.ide.editor.ui,
  org.eclipse.glsp.ide.editor.utils
-Import-Package: javax.servlet;version="[4.0.0,5.0.0)",
+Import-Package: javax.annotation;version="[1.3.5,2.0.0)",
+ javax.servlet;version="[4.0.0,5.0.0)",
  javax.websocket;version="1.1.0",
  javax.websocket.server;version="1.1.0",
  org.apache.logging.log4j;version="2.19.0",
