+++ server/plugins/org.eclipse.glsp.ide.editor/src/org/eclipse/glsp/ide/editor/handlers/IdeActionHandler.java
@@ -25,13 +25,12 @@ import org.eclipse.core.commands.AbstractHandler;
 import org.eclipse.core.commands.ExecutionEvent;
 import org.eclipse.core.commands.ExecutionException;
 import org.eclipse.e4.core.contexts.IEclipseContext;
-import org.eclipse.glsp.ide.editor.ui.GLSPDiagramEditor;
+import org.eclipse.glsp.ide.editor.ui.GLSPDiagramComposite;
 import org.eclipse.glsp.server.actions.Action;
 import org.eclipse.glsp.server.actions.ActionDispatcher;
 import org.eclipse.glsp.server.actions.ActionMessage;
 import org.eclipse.glsp.server.model.GModelState;
 import org.eclipse.glsp.server.protocol.GLSPClient;
-import org.eclipse.ui.IEditorPart;
 import org.eclipse.ui.IWorkbenchPartSite;
 import org.eclipse.ui.handlers.HandlerUtil;
 
@@ -55,7 +54,7 @@ public abstract class IdeActionHandler extends AbstractHandler {
 
    protected void dispatchMessage(final IEclipseContext context, final Action action) {
       ActionDispatcher dispatcher = context.get(ActionDispatcher.class);
-      String clientId = (String) context.get(GLSPDiagramEditor.GLSP_CLIENT_ID);
+      String clientId = (String) context.get(GLSPDiagramComposite.GLSP_CLIENT_ID);
       // Note: GLSPClient is not available at the moment, as we don't have a way to track the
       // client connection lifecycle in the Eclipse Integration yet.
       Optional<GLSPClient> client = Optional.ofNullable(context.get(GLSPClient.class));
@@ -64,9 +63,9 @@ public abstract class IdeActionHandler extends AbstractHandler {
    }
 
    protected <T> Optional<T> getInstance(final IEclipseContext context, final Class<T> type) {
-      IEditorPart editor = context.get(IEditorPart.class);
-      return editor instanceof GLSPDiagramEditor
-         ? Optional.ofNullable(((GLSPDiagramEditor) editor).getInjector().getInstance(type))
+      GLSPDiagramComposite diagramComposite = context.get(GLSPDiagramComposite.class);
+      return diagramComposite != null
+         ? Optional.ofNullable(diagramComposite.getInjector().getInstance(type))
          : Optional.empty();
    }
